Partial fix
In the previous commit, I hadn't realised that javax.json is not
available on the client side. This resulted in a maven build error. So
here I have replaced it with JSONObject. Unfortunately, I forgot the
meaning of the word "shared" in the package containing JobOption and
JobType! This results in failure at deployment time.
I think that the best fix (without wheeling in yet more JSON tech) is to
pull back the JSONification of the shared model types to the class
(JobManager?) where it is needed.