submitBatch: remove job if anything goes wrong
Any exception raised during preparation for submission is caught; the
job removed from the persistent store, and the exception is re-thrown.
If a JsonException is raised when parsing the response from the batch
connector, we check the job's JobId and Batch values; if either is null,
then the job cannot be used and so is removed from the persistent store.

Fixes #57 .