Make provenance options optional
Added new createsProvenance boolean attribute to the JobType XML;
modified JobManagementBean so that the ijpUrl and ijpJobId parameters
are only set for jobtypes with createsProvenance true. In addition, when
jobtypes are processed, only check for corresponding Applications in
ICAT for jobtypes that have createsProvenance true.

Fixes #50
TG-14 #ready-for-test-deployment
TG-21 #ready-for-test-deployment