Check session lifetime for active jobs
Record the sessionId for each job (include SQL script to add new column
to existing table). Create scheduled method (running every 15 minutes)
that looks at jobs that are (possibly) active; if a job has a sessionId,
get the remaining minutes from ICAT, and if less than 20 minutes are
left, refresh the sessionId.

This addresses the batch side of #14.