Update for icat.utils 15.1
Modified setup script and properties files to work with icat.utils
version 4.15.1. (Still to persuade my maven repo to recognise 4.16.0.)

Reordered servlets in web.xml, as setup_utils appears to overwrite
properties of the first servlet; I thought that this might be the cause
of errant behaviour that subsequently turned out to be a configuration
error.

This seems to work under Glassfish 4.0, with ICAT 4.8.0 and IDS 1.7.0;
still to test installation under Payara and with more modern ICAT/IDS
releases.